# internal/keyspan: avoid unnecessarily re-seeking keyspan iter