Avoid unnecessarily seeking the keyspan iterator in the InterleavingIter
when performing a seek on an already-positioned iterator with a current
span that encompasses the seek key.

```
name                                                 old time/op  new time/op  delta
CombinedIteratorSeek/range-key=false/batch=false-16  4.10µs ± 6%  3.81µs ± 2%   -6.94%  (p=0.000 n=10+10)
CombinedIteratorSeek/range-key=false/batch=true-16   6.72µs ± 4%  6.70µs ± 2%     ~     (p=0.796 n=9+9)
CombinedIteratorSeek/range-key=true/batch=false-16   29.7µs ± 6%   9.1µs ± 3%  -69.28%  (p=0.000 n=10+10)
CombinedIteratorSeek/range-key=true/batch=true-16    31.7µs ± 7%  12.2µs ± 1%  -61.57%  (p=0.000 n=10+9)
```

The benefit for CombinedIteratorSeek/range-key=false is noise: the
interleaving iterator is never invoked in that subbenchmark. We should
expect that there's noise in the other benchmarks as well, but the
benefits are large enough that there's a definite benefit.

Informs #1841.